chore(tests): Move SimpleSeedWallet and InMemoryWalletStorage.cs to Unit Tests project. Reference tests projects in e2e project
