+++ NArk.Tests.End2End/BoardingTests.cs
@@ -13,6 +13,7 @@ using NArk.Core.Services;
 using NArk.Core.Transformers;
 using NArk.Abstractions.Safety;
 using NArk.Safety.AsyncKeyedLock;
+using NArk.Tests.Common;
 using NArk.Tests.End2End.Common;
 using NArk.Tests.End2End.TestPersistance;
 using NArk.Transport.GrpcClient;
