+++ NArk.Tests.End2End/NArk.Tests.End2End.csproj
@@ -40,6 +40,7 @@
       <ProjectReference Include="..\NArk.Core\NArk.Core.csproj" />
       <ProjectReference Include="..\NArk.Storage.EfCore\NArk.Storage.EfCore.csproj" />
       <ProjectReference Include="..\NArk.Swaps\NArk.Swaps.csproj" />
+      <ProjectReference Include="..\NArk.Tests\NArk.Tests.csproj" />
     </ItemGroup>
 
 </Project>
