+++ NArk.Tests.End2End/Swaps/SwapManagementServiceTests.cs
@@ -21,6 +21,7 @@ using NArk.Tests.End2End.Common;
 using NArk.Tests.End2End.Core;
 using NArk.Tests.End2End.TestPersistance;
 using NArk.Core.Transformers;
+using NArk.Tests.Common;
 using NBitcoin;
 using DefaultCoinSelector = NArk.Core.CoinSelector.DefaultCoinSelector;
