+++ NArk.Tests.End2End/UnilateralExitTests.cs
@@ -23,6 +23,7 @@ using NArk.Core.Transformers;
 using NArk.Safety.AsyncKeyedLock;
 using NArk.Storage.EfCore.Hosting;
 using NArk.Storage.EfCore.Storage;
+using NArk.Tests.Common;
 using NArk.Tests.End2End.Common;
 using NArk.Tests.End2End.TestPersistance;
 using NArk.Transport.GrpcClient;
